requirements.txt
django==2.2
django-environ
psycopg2
djangorestframework
django-rest-swagger
corsheaders